openkore-ai/ai-service/src/autonomous_healing/requirements.txt
# Autonomous Self-Healing System Dependencies

# CrewAI Framework
crewai>=0.28.0
crewai-tools>=0.4.0

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0

# Async and utilities
aiofiles>=23.2.1
pyyaml>=6.0.1
watchdog>=3.0.0

# Data processing
python-dateutil>=2.8.2

# Database
aiosqlite>=0.19.0

# Monitoring
psutil>=5.9.0

# Logging
colorlog>=6.8.0

# Optional: Local LLM support
# ollama>=0.1.0
